- Added activity sequence searchable module - Rubocop fixes
