+++ app/views/api/activity_sequences/show.json.jbuilder
